+++ app/scripts/controllers/search-controller.js
@@ -84,9 +84,8 @@ angular.module('lizard-nxt')
       // therefore no need to call $rootScope.
       $scope.$broadcast('clean');
       $scope.box.query = null;
-      $scope.box.type = 'area';
+      $scope.box.type = 'point';
       $scope.box.showCards = true;
-      //console.log($scope);
     };
 
   $scope.showDetails = function (obj) {
@@ -96,7 +95,7 @@ angular.module('lizard-nxt')
       var bounds = new L.LatLngBounds(southWest, northEast);
       $scope.mapState.panZoom = bounds;
     } else {
-      if (JS_DEBUG) {
+      if (window.JS_DEBUG) {
         console.error('Oops, no boundingbox on this result - TODO: show a proper message instead of this console error...');
       }
     }