+++ CHANGES.rst
@@ -4,6 +4,22 @@ Changelog of lizard-nxt client
 Unreleased
 ----------
 
+- Fix point data for interval and ratio data.
+
+
+Release 0.2b9 (2014-10-30)
+--------------------------
+
+- Fix for double data with tiled vector layers.
+
+- Layer logic lives in its own class.
+
+- Double click performs rescale.
+
+- Clicks in the data menu wait 300 ms for a doubleclick.
+
+- Layergroups have an opacity slider that sets opacity on all leaflet layers.
+
 - Point and line give visual feedback on the map when loading and recieving data.
 
 - Images for structures added to omnibox.